Support ZisK on macOS

+++ emulator-asm/asm-runner/src/asm_min_traces_runner.rs
@@ -7,7 +7,7 @@ use named_sem::NamedSemaphore;
 use rayon::ThreadPoolBuilder;
 use zisk_common::{ChunkId, EmuTrace};
 
-use std::ffi::{c_void, CString};
+use std::ffi::{c_uint, c_void, CString};
 use std::fmt::Debug;
 use std::path::Path;
 use std::process::{self, Command};
@@ -354,7 +354,11 @@ impl AsmRunnerMT {
         unsafe { shm_unlink(shmem_input_name_ptr) };
 
         let shm_fd = unsafe {
-            shm_open(shmem_input_name_ptr, libc::O_RDWR | O_CREAT, S_IRUSR | S_IWUSR | S_IXUSR)
+            shm_open(
+                shmem_input_name_ptr,
+                libc::O_RDWR | O_CREAT,
+                (S_IRUSR | S_IWUSR | S_IXUSR) as c_uint,
+            )
         };
         Self::check_shm_open(shm_fd, shmem_input_name_ptr);
 
@@ -390,8 +394,9 @@ impl AsmRunnerMT {
         let shmem_output_name = CString::new(shmem_output_name).expect("CString::new failed");
         let shmem_output_name_ptr = shmem_output_name.as_ptr();
 
-        let shm_fd =
-            unsafe { shm_open(shmem_output_name_ptr, libc::O_RDONLY, S_IRUSR | S_IWUSR | S_IXUSR) };
+        let shm_fd = unsafe {
+            shm_open(shmem_output_name_ptr, libc::O_RDONLY, (S_IRUSR | S_IWUSR | S_IXUSR) as c_uint)
+        };
 
         Self::check_shm_open(shm_fd, shmem_output_name_ptr);
